Added blue print for piece and move validation
Piece is going to be upper level "parent" class ish.
Moves is going to be used later on for move validation.
Only works with specific x, y coordinates right now,
does not take into account if pieces are in the way.

Tested move validation by just passing in arguments
and looked at output.