Can render knight on board
Made it so that right now, we can render a knigt on the board.
Removed modular from package.json because it made it so that require
did not work in server.js.
Added modular inside of index.html because we can just have it here
for export to work.
Put board and img inside of public to be inline with everything else
and because since we are using the middleware for public, imgs would not
work if it was outside.