Changed players from socket id -> cookie id
Removed usage of socket id from game server + client and replaced it
with user account cookie to remember which user is playing the game. Currently,
there is a bug with rejoin where piece.getPlayer() is causing issues.

Changed the way pieces are displayed in 2d array. Instead of
the complex object, it is serialized to be a simple object so it
can be sent to server.

The piece can then be recreated using deserialized.