Fix pawn jump bug and double move
Because of the way we were doing first move
being 2 steps, the serialization messed it up and made
it so that every move would let the pawn go 2 spaces instead
of 1. Fixed it with a different implementation.

Fixed bug in which Pawn could jump over a piece if
it were directly infront of it when it was at its initial position