Added check and checkmate conditions
Pieces now can't move if it puts the king in check.
Pieces now also have to move in a way to get
king out of check.

Added checkmate condition but haven't implemented it
to delete player from turn pool and what happens when
there is only one player left.

Mostly here to have a clean slate of version that works
so when I play around with it nothing breaks.